refactor: 파싱 책임 분리
