+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -1,106 +1,17 @@
 package org.apache.coyote.http11;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.util.HashMap;
 import java.util.Map;
 
 public record HttpRequest(
-        String method,
-        String requestTarget,
-        String version,
-        Map<String, String> headers,
-        String body
+    String method,
+    String path,
+    String version,
+    Map<String, String> headers,
+    String body,
+    Map<String, String> queryParams
 ) {
     
-    public static HttpRequest from(BufferedReader bufferedReader) throws IOException {
-        String startLine = readRequestLine(bufferedReader);
-        
-        String[] parts = startLine.split(" ", 3);
-        if (parts.length != 3) {
-            throw new IllegalArgumentException("Invalid HTTP request format: " + startLine);
-        }
-        
-        String method = parts[0];
-        String requestTarget = parts[1];
-        String version = parts[2];
-        
-        Map<String, String> headers = parseHeaders(bufferedReader);
-        String body = parseBody(bufferedReader, headers);
-        
-        return new HttpRequest(method, requestTarget, version, headers, body);
-    }
-
-    private static String readRequestLine(BufferedReader bufferedReader) throws IOException {
-        String startLine = bufferedReader.readLine();
-        if (startLine == null || startLine.trim().isEmpty()) {
-            throw new IllegalArgumentException("Invalid request: empty start line");
-        }
-        return startLine;
-    }
-
-    private static Map<String, String> parseHeaders(BufferedReader bufferedReader) throws IOException {
-        Map<String, String> headers = new HashMap<>();
-        String line;
-        while ((line = bufferedReader.readLine()) != null && !line.isEmpty()) {
-            int colonIndex = line.indexOf(":");
-            if (colonIndex > 0) {
-                String key = line.substring(0, colonIndex).trim();
-                String value = line.substring(colonIndex + 1).trim();
-                headers.put(key, value);
-            }
-        }
-        return headers;
-    }
-
-    private static String parseBody(BufferedReader bufferedReader, Map<String, String> headers) throws IOException {
-        String contentLengthHeader = headers.get("Content-Length");
-        if (contentLengthHeader == null || contentLengthHeader.isEmpty()) {
-            return "";
-        }
-
-        try {
-            int contentLength = Integer.parseInt(contentLengthHeader);
-            if (contentLength <= 0) {
-                return "";
-            }
-
-            char[] buffer = new char[contentLength];
-            int totalRead = 0;
-
-            // Content-Length만큼 정확히 읽기
-            while (totalRead < contentLength) {
-                int read = bufferedReader.read(buffer, totalRead, contentLength - totalRead);
-                if (read == -1) {
-                    break; // EOF 도달
-                }
-                totalRead += read;
-            }
-
-            return new String(buffer, 0, totalRead);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("Invalid Content-Length header: " + contentLengthHeader);
-        }
-    }
-
-    public Map<String, String> parseQueryString() {
-        Map<String, String> params = new HashMap<>();
-
-        int queryStart = requestTarget.indexOf("?");
-        if (queryStart == -1) {
-            return params;
-        }
-
-        String queryString = requestTarget.substring(queryStart + 1);
-        String[] pairs = queryString.split("&");
-
-        for (String pair : pairs) {
-            String[] keyValue = pair.split("=");
-            if (keyValue.length == 2) {
-                params.put(keyValue[0], keyValue[1]);
-            }
-        }
-
-        return params;
+    public String getQueryParam(String name) {
+        return queryParams.get(name);
     }
 }
